定时任务基本方法

+++ src/main/java/com/azada/job/schedule/TestSchedule.java
@@ -17,9 +17,9 @@ public class TestSchedule extends BaseSchedule {
     }
 
     @Override
-    public List<Integer> getIdList() {
-        Integer[] arr = {1,2,3,4,5,6,7,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24};
-        List<Integer> testList = Arrays.asList(arr);
+    public List<Long> getIdList() {
+        Long[] arr = {1L,2L,3L,4L,5L,6L,7L,9L,10L,11L,12L,13L,14L,15L,16L,17L,18L,19L,20L,21L,22L,23L,24L};
+        List<Long> testList = Arrays.asList(arr);
         return testList;
     }
